requirements.txt
#
# This file is autogenerated by pip-compile
# Make changes in requirements.in, then run this to update:
#
#    pip-compile requirements.in
#
django-cors-headers==1.1.0
django-enumfield==1.3b2
django-extensions==1.6.1
django-polymorphic==0.8.1
django-redis==4.3.0
django==1.9.1
elasticsearch-dsl==0.0.9
elasticsearch==2.2.0
factory-boy==2.6.0
fake-factory==0.5.3
hiredis==0.2.0
psycopg2==2.6.1
python-dateutil==2.4.2    # via elasticsearch-dsl
pyyaml==3.11
redis==2.10.5
six==1.10.0
urllib3==1.14             # via elasticsearch

# The following packages are commented out because they are
# considered to be unsafe in a requirements file:
# setuptools==19.2          # via django-polymorphic
